+++ mobile/android/android-components/components/browser/thumbnails/src/test/java/mozilla/components/browser/thumbnails/ThumbnailsMiddlewareTest.kt
@@ -12,6 +12,7 @@ import mozilla.components.browser.state.state.BrowserState
 import mozilla.components.browser.state.state.createTab
 import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.browser.thumbnails.storage.ThumbnailStorage
+import mozilla.components.support.images.ImageRequest
 import mozilla.components.support.test.ext.joinBlocking
 import mozilla.components.support.test.mock
 import org.junit.Test
@@ -23,7 +24,7 @@ class ThumbnailsMiddlewareTest {
 
     @Test
     fun `thumbnail storage stores the provided thumbnail on update thumbnail action`() {
-        val sessionIdOrUrl = "test-tab1"
+        val request = ImageRequest("test-tab1")
         val tab = createTab("https://www.mozilla.org", id = "test-tab1")
         val thumbnailStorage: ThumbnailStorage = mock()
         val store = BrowserStore(
@@ -32,8 +33,8 @@ class ThumbnailsMiddlewareTest {
         )
 
         val bitmap: Bitmap = mock()
-        store.dispatch(ContentAction.UpdateThumbnailAction(sessionIdOrUrl, bitmap)).joinBlocking()
-        verify(thumbnailStorage).saveThumbnail(sessionIdOrUrl, bitmap)
+        store.dispatch(ContentAction.UpdateThumbnailAction(request.id, bitmap)).joinBlocking()
+        verify(thumbnailStorage).saveThumbnail(request, bitmap)
     }
 
     @Test